flaskapp/requirements.txt
Flask
anthropic
gunicorn
openai
pydantic-core
Flask-Login
Flask-WTF
Flask-Bcrypt
Flask-SQLAlchemy
PyMySQL
Flask-Migrate
stripe
email-validator
itsdangerous>=2.1
Werkzeug>=2.3
Jinja2>=3.1

Flask-Cors
pandas
numpy

cryptography>=42.0
flask-limiter>=3.7
redis>=5.0

google-ads>=25.0
google-api-python-client==2.136.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1



google-auth>=2.29,<3
google-auth-oauthlib>=1.2,<2
google-api-python-client>=2.133,<3
google-analytics-data==0.18.19
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
